All UI and the inventory system is held within the package. (MAKE A COPY OF THE PROJECT)
All scripts are held within this file. Before importing the file into the original project, CREATE A COPY! 
Some objects may be overridden by the items in the package!
I had an issue where I had the UI deleted when I merged the Inventory System with the UI system, don't do the same!